+++ packages/eslint-plugin/src/rules/no-dynamic-translation-ids.ts
@@ -3,6 +3,7 @@ import { translationCallTracker } from "../common-trackers";
 import { capturedRoot } from "../tracker";
 
 type MessageIds = "no-dynamic-keys";
+type Options = [];
 
 export const meta: TSESLint.RuleMetaData<MessageIds> = {
   type: "problem",
@@ -17,8 +18,10 @@ export const meta: TSESLint.RuleMetaData<MessageIds> = {
   schema: {},
 };
 
+export const defaultOptions: Options = [];
+
 export function create(
-  context: Readonly<TSESLint.RuleContext<MessageIds, []>>
+  context: Readonly<TSESLint.RuleContext<MessageIds, Options>>
 ): TSESLint.RuleListener {
   const tracker = translationCallTracker();
   tracker.listen("translation", (_node, captured) => {