+++ packages/eslint-plugin/src/rules/no-unused-translation-ids.ts
@@ -4,6 +4,7 @@ import { catalogTracker, getCatalogData } from "../common-trackers";
 import { queryUsedTranslationIds } from "../used-ids";
 
 type MessageIds = "unused-translation-id";
+type Options = [];
 
 export const meta: TSESLint.RuleMetaData<MessageIds> = {
   type: "suggestion",
@@ -19,8 +20,10 @@ export const meta: TSESLint.RuleMetaData<MessageIds> = {
   schema: {},
 };
 
+export const defaultOptions: Options = [];
+
 export function create(
-  context: Readonly<TSESLint.RuleContext<MessageIds, []>>
+  context: Readonly<TSESLint.RuleContext<MessageIds, Options>>
 ): TSESLint.RuleListener {
   const tracker = catalogTracker();
   tracker.listen('new import("@hi18n/core").Catalog()', (node, captured) => {